Phase K Wave 12: kill the sizeof(long) bug class at the AddData<T> root
Linux x86_64 sizeof(long) is 8B; Win32 (and the wire format) is 4B. Every
previous Wave landed on this same bug — Waves 7, 9, 11 each surfaced one
or two more cases — because the bulk of the emissions go through the
PacketMethods.h AddData<T>(packet, mydata, index) template and the matching
AuxBase::AddData(buffer, data, index) overloads. A grep for *((long *) ...)
casts can't surface those template instantiations.

Root-cause fix:
  - server/src/PacketMethods.h: add template <> AddData<long> and
    AddData<unsigned long> specializations that force 4-byte emission via
    int32_t / uint32_t casts. Fixes hundreds of latent
    AddData(pptr, obj->GameID(), index)-style sites in one stroke.
    Verified safe against AccountManager.cpp's account_builder.AddData(
    field, value) — that's a different member-function overload on
    mysqlplus and is unaffected.
  - server/src/AuxClasses/AuxBase.h: add the matching (long, long &) and
    (unsigned long, long &) overloads so the AUX path is wire-correct too.

Manual sweep alongside the template fix — ~34 direct *((long *) ...) cast
sites that aren't reachable through the template:
  - server/src/PlayerConnection.cpp (14): SendDataFileToClient,
    SendAuxNameResource, SetResourceDrainLevel, SetHuskDrainLevel,
    SendProspectAUX cases 0+3, SendAttackerUpdates, SendResourceContentsAUX
  - server/src/SaveManager.cpp (2): HandlePetition inbound GameID +
    ProblemType — were reading via long* off a 4B wire slot
  - server/src/PlayerSaves.cpp (1): pos_data sector slot at offset 28
  - server/src/PlayerClass.cpp (2): subparts wire overwrite, SendToRangeList
    player_id 4B compare
  - server/src/ServerManager.cpp (1): host->h_addr_list[0] IPv4 read
  - server/src/UDPConnection.cpp (1): host->h_addr_list[0] IPv4 read
  - proxy/ServerManager.cpp (1): host->h_addr_list[0] IPv4 read
  - proxy/UDPProxyMVAS.cpp (2): TerminateClient + ToggleSendFrequency
  - proxy/ClientToSectorServer.cpp (7): QueueAuxNameSignature,
    QueueNavigation, QueueResourceName, QueueContrails (2),
    QueueProspectAUX cases 0+3
  - proxy/UDPClient.cpp (1): SignalWrongVersion (Win32-only, parity)
  - login-server/Net7SSL/UDPClient.cpp (2): g_MaxPlayerCount + g_PlayerCount

Pushback recorded in plans/99-decisions-log.md: literal global s/long/int32_t/
was wrong — internal counters / loop indices / time deltas / queue API
parameters legitimately want 64-bit on Linux. The substantive request is
"every long that touches the wire becomes int32_t". Done.

Server integrity: every change is a *tightening* toward Win32 wire fidelity
(canonical 4B). No behaviour relaxation. No security posture change. No
new input acceptance. Pure preservation correctness.

Coverage: 22/207 unchanged — this is bug-class elimination, not new opcode
coverage. The next Wave 13 work (0x002C ACTION / 0x009B WARP / 0x0014 MOVE)
will inherit a now-clean wire substrate.

Verification: full 43/43 cli-integration green in 2.3 min from a fresh
`docker compose down -v` rebuild against the patched server.

+++ login-server/Net7SSL/UDPClient.cpp
@@ -156,12 +156,15 @@ void UDPClient::RecvThread()
 				switch (opcode)
 				{
 				case ENB_OPCODE_4001_SSL_REGISTER_S_SSL:
-					g_MaxPlayerCount = *((long *) &msg[0]);
+					// Phase K Wave 12: inbound SSL register payload — count is
+					// a 4B wire value, not Linux's 8B `long`.
+					g_MaxPlayerCount = *((int32_t *) &msg[0]);
                     g_LoggedIn = true;
 					break;
 
 				case ENB_OPCODE_4002_SSL_PLAYERCOUNT:
-					g_PlayerCount = *((long *) &msg[0]);
+					// Phase K Wave 12: same as above.
+					g_PlayerCount = *((int32_t *) &msg[0]);
 					break;
 
 				case ENB_OPCODE_4004_SSL_AVATARCONFIRM_S_SSL:
